+++ packages/vui-wysiwyg/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @fastkit/vui-wysiwyg
 
+## 6.0.0
+
+### Minor Changes
+
+- Changed the name representation from `FormControl` to `FormNodeWrapper`. This release includes breaking changes.
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.18.0
+
 ## 5.0.0
 
 ### Patch Changes